requirements.txt
tqdm
numpy
grequests
fake_useragent
emoji
discord.py
beautifulsoup4